tune the heuristic and reorganize its code
